Cleaning up conventions and integer types

+++ lib/fvm/bytecode.c
@@ -3,7 +3,7 @@
 
 #include <fvm/bytecode.h>
 
-size_t fvm_bytecode_arity(fvm_bytecode op) {
+uint32_t fvm_bytecode_arity(fvm_bytecode op) {
 	switch (op) {
 	case FVMI_NOP: return 0;
 	case FVMI_PUSH8: return 1;
@@ -20,7 +20,7 @@ size_t fvm_bytecode_arity(fvm_bytecode op) {
 	}
 }
 
-size_t fvm_bytecode_sarity(fvm_bytecode op) {
+uint32_t fvm_bytecode_sarity(fvm_bytecode op) {
 	switch (op) {
 	case FVMI_NOP: return 0;
 	case FVMI_PUSH8: return 0;
